+++ tf_encrypted/protocol/aby3/model/logistic_regression.py
@@ -42,11 +42,15 @@ class LogisticRegression:
             fit_batch_op = self.backward(x, dy, self.init_learning_rate)
             return fit_batch_op
 
-    def fit(self, sess, x, y, num_batches):
+    def fit(self, sess, x, y, num_iters):
         fit_batch_op = self.fit_batch(x, y)
-        for batch in range(num_batches):
+        for batch in range(num_iters):
             sess.run(fit_batch_op, tag="fit-batch")
 
+    """
+    Revealing and printing loss in the middle of training should only be used for debugging.
+    It might leak information in a normal run.
+    """
     def loss(self, sess, x, y, player_name):
         def print_loss(y_hat, y):
             with tf.name_scope("print-loss"):
